CI: Add step to check minimal-versions (#109)
* CI: Add step to check minimal-versions

#### Problem

All of the dependencies are now relaxed and even bumped down to their
lowest possible number, but we don't check to make sure that all of the
packages can even build with the lowest declared dependencies.

#### Summary of changes

Use the excellent
[`cargo-minimal-versions`](https://github.com/taiki-e/cargo-minimal-versions)
tool to do exactly that.

It goes through each package in the workspace and tries to check it
against its lowest declared dependencies.

The tool discovered a lot of instances where we declared a dependency
that was too low, so I bumped those up.

And finally, add the check to CI.

* Separate tool installation