pubkey: Don't always mark variable as mut (#108)
#### Problem

If the `std` feature is disabled on `solana-pubkey`, there's currently a
warning in the build due to a variable that doesn't need to be marked as
`mut`.

#### Summary of changes

Gate marking the variable as `mut` depending on the feature / target.
Also, we already pull in the std crate earlier if the `std` feature is
enabled, so we don't need to do it twice.